# Conflicts:
#	docs/en/sql-reference/window-functions/index.md
#	src/Analyzer/WindowNode.cpp
#	src/Common/QueryFuzzer.cpp
#	src/Processors/Transforms/WindowTransform.cpp

+++ src/Common/FieldVisitorScale.cpp
@@ -8,11 +8,11 @@ namespace ErrorCodes
     extern const int LOGICAL_ERROR;
 }
 
-FieldVisitorScale::FieldVisitorScale(Int32 rhs_) : rhs(rhs_) {}
+FieldVisitorScale::FieldVisitorScale(Int64 rhs_) : rhs(rhs_) {}
 
 void FieldVisitorScale::operator() (Int64 & x) const { x *= rhs; }
 void FieldVisitorScale::operator() (UInt64 & x) const { x *= rhs; }
-void FieldVisitorScale::operator() (Float64 & x) const { x *= rhs; }
+void FieldVisitorScale::operator() (Float64 & x) const { x *= static_cast<Float64>(rhs); }
 void FieldVisitorScale::operator() (Null &) const { /*Do not scale anything*/ }
 
 void FieldVisitorScale::operator() (String &) const { throw Exception(ErrorCodes::LOGICAL_ERROR, "Cannot scale Strings"); }
